Tests now trigger scheduled notifications, & other fixes
Summary:
Tests now trigger scheduled notifications
(no real way to test them, the tests don't even last that long)
Minor bugfixes and refactoring
The java plugin can now be stopped safely, so that's re-enabled in the c# side of the plugin.
Renaming the playfab .aar to match the output of the java project.

Test Plan: Jenker! + a bit of manual

Reviewers: Marco.Williams

Maniphest Tasks: T2999

Differential Revision: https://phab.playfabdev.com/D3818