Push Plugin Updated & Facebook fixed
Summary:
Delete and update Facebook (project seems to be committed in a busted state because of a mix of original and generated Facebook files, with no easy way to fix it)
Delete and replace Old PlayFab push plugin with FCM (the goal)
Currently doesn't compile because the old code is looking for the PlayFab push plugin, so I have to migrate it from the other branch.
Ignoring the Firebase generated files, and other project build fixes.

Test Plan: Manual run with manual push notification

Reviewers: Marco.Williams

Maniphest Tasks: T3328

Differential Revision: https://phabricator.playfab.com/D4691